* dev_new:
  [Android] 12.3.0
  [Android] 广告更新
  [Android] 12.2.3 添加insaitlib
  [Android] 12.1.8
  [Feather] 12.1.7  广告sdk更新 激励广告设置时间8s加载
  [Feather] 12.1.6
  [Feather] 短剧、小视频
  [fix] 添加短剧
  [Android] 搜索修改
  [Feather] 闪退修改
  [Feather] 12.1.4
  [Feather] 城市搜索修改
  [Feather] 12.1.2 定位修改
  [Android] 12.1.1
  [Android] 忽略

# Conflicts:
#	app/build.gradle
#	app/src/main/AndroidManifest.xml
#	app/src/main/java/com/goodtech/tq/activity/BaseActivity.java
#	app/src/main/java/com/goodtech/tq/activity/MainActivity.java
#	app/src/main/java/com/goodtech/tq/activity/SettingActivity.java
#	app/src/main/java/com/goodtech/tq/app/BaseApp.java
#	app/src/main/java/com/goodtech/tq/fragment/WeatherFragment.java
#	app/src/main/java/com/goodtech/tq/modules/others/video/fragment/VideoPagerFragment.java

+++ app/src/main/java/com/goodtech/tq/modules/citySearch/viewholder/CityHolder.java
@@ -123,32 +123,28 @@ public class CityHolder extends AbstractDraggableItemViewHolder implements View.
     @SuppressLint("NonConstantResourceId")
     @Override
     public void onClick(View v) {
-        switch (v.getId()) {
-            case R.id.img_delete:
-                showDeleteAnim();
-                mContainer.setOnClickListener(this);
+        if (v.getId() == R.id.img_delete) {
+            showDeleteAnim();
+            mContainer.setOnClickListener(this);
+            if (mListener != null) {
+                mListener.onShowDelete(this);
+            }
+        } else if (v.getId() == R.id.container) {
+            if (isEdit) {
+                hideDeleteAnim();
+                mContainer.setOnClickListener(null);
                 if (mListener != null) {
-                    mListener.onShowDelete(this);
+                    mListener.onShowDelete(null);
                 }
-                break;
-            case R.id.container:
-                if (isEdit) {
-                    hideDeleteAnim();
-                    mContainer.setOnClickListener(null);
-                    if (mListener != null) {
-                        mListener.onShowDelete(null);
-                    }
-                } else {
-                    if (mListener != null) {
-                        mListener.onItemClick(v, getAdapterPosition(), mCityMode);
-                    }
-                }
-                break;
-            case R.id.btn_delete:
+            } else {
                 if (mListener != null) {
-                    mListener.onDeleteCity(getAdapterPosition(), mCityMode);
+                    mListener.onItemClick(v, getAdapterPosition(), mCityMode);
                 }
-                break;
+            }
+        } else if (v.getId() == R.id.btn_delete) {
+            if (mListener != null) {
+                mListener.onDeleteCity(getAdapterPosition(), mCityMode);
+            }
         }
     }
